remove magic numbers
